fetching child pages and attachments
