+++ packages/integration/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@azure/identity": "^4.0.0",
     "@azure/storage-blob": "^12.5.0",
+    "@backstage/backend-defaults": "workspace:^",
     "@backstage/config": "workspace:^",
     "@backstage/errors": "workspace:^",
     "@octokit/auth-app": "^4.0.0",
@@ -48,7 +49,8 @@
     "git-url-parse": "^15.0.0",
     "js-yaml": "^4.0.0",
     "lodash": "^4.17.21",
-    "luxon": "^3.0.0"
+    "luxon": "^3.0.0",
+    "node-html-markdown": "^1.3.0"
   },
   "devDependencies": {
     "@backstage/cli": "workspace:^",
